requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.40
tavily-python==0.3.3
yfinance==0.2.18
polygon-api-client==1.12.3
dash==2.14.1
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
python-dateutil==2.8.2
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.3
statsmodels==0.14.0
openpyxl==3.1.2
xlsxwriter==3.1.9
tiktoken>=0.5.1

